requirements.txt
torch==2.7.1
numpy==1.26.4
torchvision==0.22.1
transformers==4.57.6
huggingface-hub
peft==0.17.1
accelerate==1.10.1
datasets==4.1.1
scipy
tqdm
wandb
pillow
wrapt
py-cpuinfo
hjson
scikit-learn
scikit-image
qwen_vl_utils
ray
timm
sentencepiece
opencv-python
decord
hnswlib
einops
deepspeed==0.17.6
numba
